useStatefulSet switch in revad for stateful workloads
This is specially useful when running into sharded scenarios as
e.g. different storage providers to map users with different
rules. Ref. https://github.com/cs3org/reva/pull/1142

This commit contains a preliminary version, where the "real"
uniqueness of each replica in the set (in terms of config) is yet
to be implemented. First version will simply rely on the hostname
as selector inside the 'configFiles' value for a user-provided
toml file.